+++ src/components/dashboard/containers/Tables/AdUnitsTable.js
@@ -4,7 +4,7 @@ import classnames from 'classnames'
 import { Tooltip, IconButton } from '@material-ui/core'
 import { Visibility } from '@material-ui/icons'
 import Img from 'components/common/img/Img'
-import MUIDataTableEnchanced from 'components/dashboard/containers/Tables/MUIDataTableEnchanced'
+import MUIDataTableEnhanced from 'components/dashboard/containers/Tables/MUIDataTableEnhanced'
 import { withReactRouterLink } from 'components/common/rr_hoc/RRHoc'
 import { t, selectAdUnits, selectSide, selectNewCampaign } from 'selectors'
 import { makeStyles } from '@material-ui/core/styles'
@@ -155,7 +155,7 @@ function AdUnitsTable(props) {
 		},
 	]
 	return (
-		<MUIDataTableEnchanced
+		<MUIDataTableEnhanced
 			title={campaignUnits ? t('CAMPAIGN_AD_UNITS') : t('ALL_UNITS')}
 			data={data}
 			columns={columns}
